software/backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.0.1
boto3==1.37.29
botocore==1.37.29
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.11
h11==0.14.0
idna==3.10
iniconfig==2.1.0
jmespath==1.0.1
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.10.6
pydantic_core==2.27.2
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
rsa==4.9
s3transfer==0.11.4
secret==0.8
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
tabulate==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
